+++ src/components/add-contract/CompilerDetails.tsx
